+++ src/ImageProcessing.jl
@@ -2,8 +2,8 @@ module ImageProcessing
 
 export
     IndexBox,
-    Point,
     OnlineSum,
+    Point,
     center_of_gravity,
     hard_thresholder,
     nearest,
@@ -28,8 +28,6 @@ using StructuredArrays
 using TypeUtils
 
 include("macros.jl")
-@public to_dim to_size to_axis to_axes
-
 include("compat.jl")
 include("types.jl")
 include("utils.jl")